Enforce common dependency configuration setup
